# openapi.yaml
openapi: 3.0.3
info:
  title: Open Mic API
  version: 1.0.0
  contact:
    email: info@openmics.org
  description:
    Canonical API contract. All paths are relative to the /api base URL.
    The public media deep-link page (`/media/:mediaId`) is intentionally not part of
    this document — like the `/@:handle` vanity resolver it is served outside `/api`
    as server-rendered SPA HTML with Open Graph tags, not as a JSON operation.
servers:
  - url: https://api.openmics.org/api
    description: Production
...
paths:
...
  /open-mics:
    get:
      tags:
        - Open Mics
      operationId: listOpenMics
      description: Returns a paginated list of open-mic series matching the supplied search, location, activity, rating, registration, and sort filters.
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/City"
        - $ref: "#/components/parameters/Activity"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/Near"
        - $ref: "#/components/parameters/Radius"
        - $ref: "#/components/parameters/MinRating"
        - $ref: "#/components/parameters/RegistrationMode"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/OpenMicListResponse"
    post:
      tags:
        - Open Mics
      operationId: createOpenMic
      description: Creates an open-mic series using the supplied details.
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/OpenMicCreate"
      responses:
        "201":
          $ref: "#/components/responses/OpenMicResponse"
        "403":
          $ref: "#/components/responses/PlanLimitExceeded"
        "409":
          $ref: "#/components/responses/HandleConflict"
  /me/open-mics:
    get:
      tags:
        - Open Mics
      operationId: listMyOpenMics
      description: Lists every non-deleted series owned by the given profile, regardless of status (including draft/paused/ended). Unlike GET /open-mics (the public directory, which always excludes draft/ended), this always requires the caller to own the profile for the organizer's own dashboard, so a just-created draft series is visible immediately.
      security:
        - bearerAuth: []
      parameters:
        - name: owner_profile_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/OpenMicListResponse"
  /open-mics/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags:
        - Open Mics
      operationId: getOpenMic
      description: Active series are publicly readable. Draft, paused and ended series are available only to their owning account or a platform admin; deleted series are unavailable. Paused series are also excluded from public discovery and media.
      responses:
        "200":
          $ref: "#/components/responses/OpenMicResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Open Mics
      operationId: deleteOpenMic
      description: Deletes the open-mic series identified by `id`.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Open mic soft-deleted
    patch:
      tags:
        - Open Mics
      operationId: updateOpenMic
      description: Partially updates the open-mic series identified by `id`.
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/OpenMicUpdate"
      responses:
        "200":
          $ref: "#/components/responses/OpenMicResponse"
...
  /open-mics/{id}/kiosk-backup-pin:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags:
        - Open Mics
      operationId: getKioskBackupPinStatus
      summary: Whether this series has a kiosk backup PIN configured
      description: Reports whether the open-mic series has a kiosk backup PIN configured without revealing the PIN.
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/KioskBackupPinStatusResponse"
    put:
      tags:
        - Open Mics
      operationId: setKioskBackupPin
      summary: Set (or replace) this series' kiosk backup PIN
      description: Sets or replaces the kiosk backup PIN for the open-mic series.
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/KioskBackupPinSet"
      responses:
        "200":
          $ref: "#/components/responses/KioskBackupPinStatusResponse"
  /open-mics/{id}/kiosk-backup-pin/reveal:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags:
        - Open Mics
      operationId: revealKioskBackupPin
      summary: Explicitly reveal this series' configured kiosk backup PIN
      description: Owner-authenticated. Returns the stored PIN only on explicit request and disables caching.
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/KioskBackupPinRevealResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/KioskBackupPinNotConfigured"
  /open-mics/{id}/kiosk-backup-pin/verify:
    post:
      tags:
        - Open Mics
      operationId: verifyKioskBackupPin
      description: Checks whether the submitted PIN matches the configured kiosk backup PIN.
      summary: Check an entered PIN against this series' configured kiosk backup PIN
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        $ref: "#/components/requestBodies/KioskBackupPinSet"
      responses:
        "200":
          $ref: "#/components/responses/KioskBackupPinVerifyResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    editSession:
      type: apiKey
      in: cookie
      name: openmic_edit_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Q:
      name: q
      in: query
      schema:
        type: string
    Country:
      name: country
      in: query
      schema:
        type: string
    City:
      name: city
      in: query
      schema:
        type: string
    Activity:
      name: activity
      in: query
      schema:
        $ref: "#/components/schemas/Activity"
    Tag:
      name: tag
      in: query
      schema:
        type: string
    Near:
      name: near
      in: query
      description: Latitude and longitude as `lat,lng`.
      schema:
        type: string
        pattern: ^-?[0-9.]+,-?[0-9.]+
    Radius:
      name: radius_km
      in: query
      schema:
        type: number
        minimum: 0
    MinRating:
      name: min_rating
      in: query
      schema:
        type: number
        minimum: 0
        maximum: 5
    RegistrationMode:
      name: registration_mode
      in: query
      schema:
        $ref: "#/components/schemas/RegistrationMode"
    Sort:
      name: sort
      in: query
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
  requestBodies:
...
    OpenMicCreate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenMicCreateRequest"
    OpenMicUpdate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenMicUpdateRequest"
...
    KioskBackupPinSet:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/KioskBackupPinRequest"
  responses:
...
    OpenMicResponse:
      description: Open mic
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenMic"
    KioskBackupPinStatusResponse:
      description: Whether a kiosk backup PIN is configured for this series
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/KioskBackupPinStatus"
    KioskBackupPinRevealResponse:
      description: Explicitly requested configured PIN. This response must not be cached.
      headers:
        Cache-Control:
          description: Prevents storage of the revealed PIN in caches.
          schema:
            type: string
            enum:
              - no-store
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/KioskBackupPinReveal"
    KioskBackupPinNotConfigured:
      description: No kiosk backup PIN is configured for this series
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    KioskBackupPinVerifyResponse:
      description: Whether the submitted PIN matches this series' configured kiosk backup PIN
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/KioskBackupPinVerifyResult"
...
    OpenMicListResponse:
      description: Open mics
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenMicPage"
...
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    HandleConflict:
      description: Handle unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HandleConflictError"
...
    PlanLimitExceeded:
      description:
        "`error.code` is `PLAN_LIMIT_EXCEEDED` with `error.details.scope` =
        `series` (max series per organizer reached), `event` (max events per series
        reached), or `event_capacity` (capacity above the plan maximum, or unlimited
        capacity, which the default plan rejects)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Id:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    Activity:
      type: string
      enum:
        - singing
        - poetry
        - jam
        - trad
        - comedy
        - storytelling
        - other
...
    OpenMicStatus:
      type: string
      enum:
        - active
        - paused
        - ended
        - draft
    RegistrationMode:
      type: string
      enum:
        - pre_only
        - on_night_only
        - both
        - external
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
              description:
                Machine-readable extras specific to `code` — validation field
                errors for VALIDATION_ERROR, or a `scope` discriminator for
                PLAN_LIMIT_EXCEEDED / MEDIA_QUOTA_EXCEEDED.
    HandleConflictError:
      allOf:
        - $ref: "#/components/schemas/ErrorEnvelope"
        - type: object
          properties:
            holder_type:
              type: string
              enum:
                - profile
                - open_mic
              nullable: true
            holder_id:
              allOf:
                - $ref: "#/components/schemas/Id"
              nullable: true
...
    OpenMic:
      type: object
      required:
        - id
        - current_handle
        - name
        - status
        - city_id
      properties:
        id:
          $ref: "#/components/schemas/Id"
        public_code:
          type: string
          pattern: "^[A-Z0-9]{10}$"
          description: Short stable public identifier for URLs. The UUID id remains the internal identifier.
        owner_profile_id:
          $ref: "#/components/schemas/Id"
        current_handle:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        tags:
          type: array
          items:
            type: string
        venue_name:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
          nullable: true
        postcode:
          type: string
          nullable: true
        city:
          type: string
        country:
          type: string
        city_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
        distance_km:
          type: number
          minimum: 0
          description: Present on geographically scoped public results.
        lat:
          type: number
          minimum: -90
          maximum: 90
          nullable: true
        lng:
          type: number
          minimum: -180
          maximum: 180
          nullable: true
        time_zone:
          type: string
          example: Europe/Dublin
        schedule_summary:
          type: string
          nullable: true
        schedule_details:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        contact_email:
          type: string
          format: email
          nullable: true
        originals_only:
          type: boolean
        amplification_available:
          type: boolean
        age_policy:
          type: string
          enum:
            - adults_only
            - children_only
            - both
        registration_mode:
          $ref: "#/components/schemas/RegistrationMode"
        external_registration_url:
          type: string
          format: uri
          nullable: true
        entry_fee_amount:
          type: number
          nullable: true
        entry_fee_currency:
          type: string
          nullable: true
        entry_fee_note:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OpenMicStatus"
        rating_avg:
          type: number
          minimum: 0
          maximum: 5
        rating_count:
          type: integer
          minimum: 0
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
    OpenMicPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OpenMic"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    OpenMicCreateRequest:
      type: object
      required:
        - name
        - venue_name
        - address_line1
        - time_zone
      properties:
        name:
          type: string
        description:
          type: string
        handle:
          type: string
        venue_name:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
        postcode:
          type: string
        city:
          type: string
        country:
          type: string
        city_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
          description: Selected catalogue city. When set to a UUID, the server validates and derives canonical city/country snapshots; null leaves free-text snapshots unlinked.
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        time_zone:
          type: string
        website:
          type: string
          format: uri
        contact_email:
          type: string
          format: email
        schedule_summary:
          type: string
        schedule_details:
          type: string
        originals_only:
          type: boolean
        amplification_available:
          type: boolean
        age_policy:
          type: string
          enum:
            - adults_only
            - children_only
            - both
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        tags:
          type: array
          items:
            type: string
        registration_mode:
          $ref: "#/components/schemas/RegistrationMode"
        external_registration_url:
          type: string
          format: uri
        entry_fee_amount:
          type: number
          minimum: 0
        entry_fee_currency:
          type: string
        entry_fee_note:
          type: string
...
    OpenMicUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        handle:
          type: string
        venue_name:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
        postcode:
          type: string
        city:
          type: string
        country:
          type: string
        city_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
          description: Omission preserves the reference; null clears it. Changing free-text city or country without a selected ID clears a stale reference.
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        time_zone:
          type: string
        website:
          type: string
          format: uri
        contact_email:
          type: string
          format: email
        schedule_summary:
          type: string
        schedule_details:
          type: string
        originals_only:
          type: boolean
        amplification_available:
          type: boolean
        age_policy:
          type: string
          enum:
            - adults_only
            - children_only
            - both
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        tags:
          type: array
          items:
            type: string
        registration_mode:
          $ref: "#/components/schemas/RegistrationMode"
        external_registration_url:
          type: string
          format: uri
        entry_fee_amount:
          type: number
          minimum: 0
        entry_fee_currency:
          type: string
        entry_fee_note:
          type: string
...
    KioskBackupPinRequest:
      type: object
      description: The organizer's kiosk backup PIN in plaintext. Routine and public reads omit it; it is returned only by the explicit owner-authenticated reveal operation.
      required:
        - pin
      properties:
        pin:
          type: string
          minLength: 4
    KioskBackupPinStatus:
      type: object
      required:
        - configured
      properties:
        configured:
          type: boolean
    KioskBackupPinReveal:
      type: object
      required:
        - pin
      properties:
        pin:
          type: string
    KioskBackupPinVerifyResult:
      type: object
      required:
        - valid
      properties:
        valid:
          type: boolean
...